Move publishRootchainMetrics to stats.go
